Convert pagination schema type to typescript
